更新 開課查詢功能
修改 開課查詢功能的傳入參數
