+++ Member/Administrator.java
@@ -11,6 +11,7 @@ public class Administrator extends Member {
 	public Administrator(String id, char[] password, String name) {
 		super(id,new String(password),name);
 	}
+
 	//列出課程學生成績
 	public Object[][] getCourseStudentScore(String year, String subjectName)  {
 		//建檔案路徑
@@ -353,7 +354,6 @@ public class Administrator extends Member {
 		// no work
 		return null;
 	}
-
 	// 檢查學生存在
 	public boolean checkStudentExist(String id) {
 		// 檢查 學生是否存在
@@ -368,7 +368,6 @@ public class Administrator extends Member {
 		}
 		return check;
 	}
-
 	// 新增學生資訊
 	public boolean addStudentInfo(String id, String name, String startSchool) {
 		try {
@@ -415,7 +414,6 @@ public class Administrator extends Member {
 		}
 		return true;
 	}
-
 	// 執行檔案刪除
 	private void deleteFile(File file) {
 		// 判斷路徑是否存在
@@ -457,7 +455,6 @@ public class Administrator extends Member {
 		}
 		return data;
 	}
-
 	// 修改學生資訊
 	public boolean changeStudentInformation(String newId, String newName, String newStartSchool) {
 		try {
@@ -519,43 +516,6 @@ public class Administrator extends Member {
 		}
 		return true;
 	}
-
-	//複製整個資料夾內容
-	private void copyFolder(String oldPath, String newPath) {
-		try {
-			(new File(newPath)).mkdirs(); // 如果資料夾不存在則建立新資料夾
-			File a = new File(oldPath);
-			String[] file = a.list();
-			File temp = null;
-			for (int i = 0; i < file.length; i++) {
-				if (oldPath.endsWith(File.separator)) {
-					temp = new File(oldPath + file[i]);
-				} else {
-					temp = new File(oldPath + File.separator + file[i]);
-				}
-
-				if (temp.isFile()) {
-					FileInputStream input = new FileInputStream(temp);
-					FileOutputStream output = new FileOutputStream(newPath + "/" + (temp.getName()).toString());
-					byte[] b = new byte[1024 * 5];
-					int len;
-					while ((len = input.read(b)) != -1) {
-						output.write(b, 0, len);
-					}
-					output.flush();
-					output.close();
-					input.close();
-				}
-				if (temp.isDirectory()) {// 如果是子資料夾
-					copyFolder(oldPath + "/" + file[i], newPath + "/" + file[i]);
-				}
-			}
-		} catch (Exception e) {
-			System.out.println("複製整個資料夾內容操作出錯");
-			e.printStackTrace();
-		}
-	}
-
 	//-------------------------------------------------------------------------------------
 	//檢查課程是否存在
 	private boolean checkSubjectExist(String year, String subject) {
@@ -712,7 +672,7 @@ public class Administrator extends Member {
 		}
 		return null;
 	} 
-	// 修改 課程資訊
+	//修改 課程資訊
 	public boolean changeSubjectInformation(String year, String subject, String newYear, String newId, String newSubject,
 			String newCredit, String newType, String newTeacher) {
 		// 課程存在
