+++ TeacherWriteScoreFrame.java
@@ -5,7 +5,6 @@ import javax.swing.table.JTableHeader;
 import javax.swing.table.TableColumn;
 
 import Member.Teacher;
-import Member.Member;
 
 import java.awt.event.*;
 import java.io.BufferedReader;
@@ -97,7 +96,7 @@ public class TeacherWriteScoreFrame extends JFrame implements ActionListener {
     		public void itemStateChanged(ItemEvent e) {
 				if(e.getStateChange() == ItemEvent.SELECTED){
 					Object[][] data = user.getCourseStudentScore((String) jcb_semester.getSelectedItem(), (String) jcb_course.getSelectedItem());
-					Member.cleanTable(tableM);
+					cleanTable(tableM);
 					for(int i = 0; i < data.length; i++){
 						tableM.addRow(data[i]);
 					}
@@ -180,5 +179,10 @@ public class TeacherWriteScoreFrame extends JFrame implements ActionListener {
             Integer.parseInt(msg);
 			return true;
 		}catch(Exception e){return false;}
-    }
+	}
+	// 清空表單method
+	public static void cleanTable(DefaultTableModel table) {
+		while (table.getRowCount() > 0)
+			table.removeRow(0);
+	}
 }
