+++ studentCourseFrame.java
@@ -5,7 +5,6 @@ import javax.swing.table.JTableHeader;
 import javax.swing.table.TableColumn;
 
 import Member.Administrator;
-import Member.Member;
 
 import java.awt.event.*;
 import java.io.File;
@@ -135,7 +134,7 @@ public class studentCourseFrame extends JFrame implements ActionListener {
 		String student = jid.getText();
         if(e.getSource() == jcb_time) {
 			Object[][] courseList = user.getStudentCourseList(selectedSemester, student);
-			Member.cleanTable(tableM);
+			cleanTable(tableM);
 			for(int i = 0; i < courseList.length; i++) {
 				tableM.addRow(courseList[i]);
 			}
@@ -166,5 +165,11 @@ public class studentCourseFrame extends JFrame implements ActionListener {
         		new removeStuCourseFrame(user, student);
         	}
         }
-    }
+	}
+	
+	// 清空表單method
+	public static void cleanTable(DefaultTableModel table) {
+		while (table.getRowCount() > 0)
+			table.removeRow(0);
+	}
 }
